+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 3.8.0 (Unpublished)
+* Add `DateTimePrinter` to print date and time objects.
 * Add `OrdinalNumberPrinter` to print ordinal numbers: 1st, 2nd, 3rd, 4th, ...
 * Add `Printer.switcher` to use different printers based on conditions.
 * Add `Printer.where` to condition the printing on the printed object.